+++ inc/features/clonable.class.php
@@ -37,6 +37,7 @@ if (!defined('GLPI_ROOT')) {
 }
 
 use CommonDBConnexity;
+use CommonDBTM;
 use Toolbox;
 
 /**
@@ -45,20 +46,49 @@ use Toolbox;
 trait Clonable {
 
    /**
-    * Get relations class to clone along with current element
+    * Get relations class to clone along with current element.
     *
     * @return CommonDBTM::class[]
     */
    abstract public function getCloneRelations() :array;
 
    /**
-    * Clone the item's relations
+    * Clean input used to clone.
     *
-    * @param $source
-    * @param $history
+    * @param array $input
+    *
+    * @return array
+    *
     * @since x.x.x
     */
-   public function cloneRelations($source, $history) {
+   private function cleanCloneInput(array $input): array {
+      $properties_to_clean = [
+         'id',
+         'date_mod',
+         'date_creation',
+         'template_name',
+         'is_template'
+      ];
+      foreach ($properties_to_clean as $property) {
+         if (array_key_exists($property, $input)) {
+            unset($input[$property]);
+         }
+      }
+
+      return $input;
+   }
+
+   /**
+    * Clone the item's relations.
+    *
+    * @param CommonDBTM $source
+    * @param bool       $history
+    *
+    * @return void
+    *
+    * @since x.x.x
+    */
+   private function cloneRelations(CommonDBTM $source, bool $history): void {
       $clone_relations = $this->getCloneRelations();
       foreach ($clone_relations as $classname) {
          if (!is_a($classname, CommonDBConnexity::class, true)) {
@@ -80,18 +110,16 @@ trait Clonable {
    }
 
    /**
-    * Prepare input datas for cloning the item
+    * Prepare input datas for cloning the item.
+    * This empty method is meant to be redefined in objects that need a specific prepareInputForClone logic.
     *
     * @since x.x.x
     *
     * @param array $input datas used to add the item
     *
     * @return array the modified $input array
-    **/
-   function prepareInputForClone($input) {
-      unset($input['id']);
-      unset($input['date_mod']);
-      unset($input['date_creation']);
+    */
+   public function prepareInputForClone($input) {
       return $input;
    }
 
@@ -101,12 +129,12 @@ trait Clonable {
     * @since x.x.x
     *
     * @param array $override_input custom input to override
-    * @param boolean $history do history log ? (true by default)
+    * @param boolean $history do history log ?
     *
     * @return integer The new ID of the clone (or false if fail)
     */
    public function clone(array $override_input = [], bool $history = true) {
-      global $DB, $CFG_GLPI;
+      global $DB;
 
       if ($DB->isSlave()) {
          return false;
@@ -116,33 +144,27 @@ trait Clonable {
       foreach ($override_input as $key => $value) {
          $input[$key] = $value;
       }
+      $input = $new_item->cleanCloneInput($input);
       $input = $new_item->prepareInputForClone($input);
-      if (isset($input['id'])) {
-         $input['_oldID'] =  $input['id'];
-         unset($input['id']);
-      }
-      unset($input['date_creation']);
-      unset($input['date_mod']);
-
-      if (isset($input['template_name'])) {
-         unset($input['template_name']);
-      }
-      if (isset($input['is_template'])) {
-         unset($input['is_template']);
-      }
 
       $input['clone'] = true;
       $newID = $new_item->add($input, [], $history);
-      // If the item needs post clone (recursive cloning for example)
-      $new_item->post_clone($this, $history);
+
+      if ($newID !== false) {
+         $new_item->cloneRelations($this, $history);
+         $new_item->post_clone($this, $history);
+      }
+
       return $newID;
    }
 
    /**
+    * Post clone logic.
+    * This empty method is meant to be redefined in objects that need a specific post_clone logic.
+    *
     * @param $source
     * @param $history
     */
    public function post_clone($source, $history) {
-      $this->cloneRelations($source, $history);
    }
 }