+++ inc/networkport.class.php
@@ -48,6 +48,8 @@ if (!defined('GLPI_ROOT')) {
 **/
 class NetworkPort extends CommonDBChild {
 
+   use Glpi\Features\Clonable;
+
    // From CommonDBChild
    static public $itemtype             = 'itemtype';
    static public $items_id             = 'items_id';
@@ -59,6 +61,9 @@ class NetworkPort extends CommonDBChild {
 
    static $rightname                   = 'networking';
 
+   public function getCloneRelations() :array {
+      return [];
+   }
 
    function getForbiddenStandardMassiveAction() {
 
@@ -223,7 +228,6 @@ class NetworkPort extends CommonDBChild {
    }
 
    function post_clone($source, $history) {
-      $this->post_cloneTrait($source, $history);
       $instantiation = $source->getInstantiation();
       if ($instantiation !== false) {
          $instantiation->fields[$instantiation->getIndexName()] = $this->getID();